chatbot/requirements.txt
numpy>=1.20.0
scikit-learn>=1.0.0
groq>=0.4.0
python-dotenv>=0.19.0
faiss-cpu==1.7.4
sentence-transformers==2.5.1
streamlit==1.32.2
pandas==2.2.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3 